requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart

opencv-python
numpy
Pillow

torch
torchvision
facenet-pytorch

pinecone
jinja2